Use long Timeout property instead of short

+++ LimaDetector/src/Factory.cpp
@@ -104,8 +104,8 @@ CtControl* ControlFactory::get_control( const std::string& detector_type)
                 (Tango::Util::instance()->get_database())->get_device_property(my_device_name, db_data);
                 std::string camera_ip;
                 db_data[0] >> camera_ip;
-                short detector_timeout = 11000;
-                db_data[1] >> detector_timeout;
+                long detector_timeout = 11000;
+                db_data[1] >> detector_timeout;                
                 long packet_size = -1;
                 db_data[2] >> packet_size;
                 my_camera_basler            = new Basler::Camera(camera_ip, packet_size);
